services/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pillow>=10.0.0
numpy>=1.26.0
torch>=2.3.0
